Initial Commit - Landing Page and Waitlist Flow
General Overview:

This change adds a react-based front end client that displays a simple
landing page and provides a "waitlist signup" flow that allows users to
input an email that will be stored for future use.

Problem:

Currently in its infancy, Voto lacks any mechanisms that provide data
on features that users are interested in and who those users are.

Solution:

This commit adds a barebones landing page for Voto and a wait list flow
that allows users to input their email to be added to a waitlist. While
this doesn't directly provide a mechanism to acquire data on user's desired
feautures, this will provide a foundation for reaching a potential user
base.