Add CNAME to publish script
Problem:

The current npm 'publish' script erases the existing CNAME file that
links the voto.web.mpbell.dev CNAME to the github pages static site.
When this is removed, users cannot access the static site at
voto.web.mpbell.dev

Solution:

This commit recreates CNAME file in the publish script pipeline. This
allows the CNAME to be available in when the static files are pushed to
the gh-pages branch, and allows the publish script to include all
necessary steps to publish the current version of the static site.