Feature - Responsive Styles
Problem:

The existing implementation is visually jarring for screen sizes <
600px and presents a myriad of usability challenges on mobile devices.

Solution:

This change aims to address the responsiveness concerns with 2 major
efforts 1) Reorganization of styles and 2) Media Queries.

Initially, this changed faced friction due to the substantial amount of
inline JSX styles. Because of this, the project's styles were
reorganized into a single `styles` object that contains all static
styles in one place. This change cleanly separates layout from logic.
This unified styles object allows for extensibility and ease of
development for future style modifications.

In order to provide responsiveness to viewport width, css media queries
are used to dynamically set the `:root`'s font size. The react styles
are then able to leverages the dynamic size with the `rem` unit. Many
of the previously pixel dimensioned quantites were converted to an
appropriate `rem` alternative. By baselineing against the `rem`, the
styles become responsive to the user's viewport